requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
supabase==2.0.2
exa-py==1.0.9
PyYAML==6.0.1
python-multipart==0.0.6